+++ lib/reportinator/types/model.rb
@@ -20,16 +20,14 @@ module Reportinator
     end
 
     def get_model_data(target)
-      parsed_target = ValueParser.parse(target)
       method_list.map do |method|
-        parse_method(parsed_target, method)
+        parse_method(target, method)
       end
     end
 
     def parse_method(target, method)
-      parsed_method = ValueParser.parse(method)
-      return parsed_method if parsed_method.instance_of?(String)
-      MethodParser.parse(target, parsed_method)
+      return method if method.instance_of?(String)
+      MethodParser.parse(target, method)
     end
   end
 end